3_SLAM_System/src/requirements.txt
opencv-contrib-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
scipy==1.11.2
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200
pyyaml==6.0.1
tqdm==4.66.1
scikit-image==0.21.0
open3d==0.18.0
pillow==10.0.1
